requirements.txt
# FastAPI Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Telegram Bot
python-telegram-bot==20.7

# MongoDB
motor==3.3.2
pymongo==4.6.0

# OpenAI
openai==1.3.7

# Environment
python-dotenv==1.0.0

# Scheduling/Async
apscheduler==3.10.4

# HTTP Client
httpx==0.25.2

